+++ app/src/main/java/com/scipath/becomeaking/model/enums/Direction.java
@@ -1,0 +1,6 @@
+package com.scipath.becomeaking.model.enums;
+
+public enum Direction {
+    Left,
+    Right
+}